crypto: inside-secure: remove unused enqueue return code
When invalidation result is handled, it enqueues the original
request which was the cause for the invalidation, so since
this invalidation result does not require completion, the
enqueue return code will not be used and thus can be dropped.

In any case the request is in the queue regardless to the
enqueue return code.

Change-Id: I2446f49d4aab97303220169bea74cc13f754b75d
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/45613
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>