arm64: defconfig: disable DMA_CMA by default
The PPv2 driver allocate lots of DMA buffers. These buffers are allocated
thanks to CMA if CONFIG_DMA_CMA is enabled, the size of the DMA pool is
configurable and too small in the default configuration.
This lead to the driver failing to probe because of this.
One solution is too disable CONFIG_DMA_CMA.
This patch might be reverted when backport patches for better solution.

Change-Id: Iac578de86fe34e40b887c17e50c4b312d5888069
Signed-off-by: Antoine Tenart <antoine.tenart@free-electrons.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52054
Reviewed-by: Victor Gu <xigu@marvell.com>
Tested-by: Victor Gu <xigu@marvell.com>