arm64: defconfig: disable DMA_CMA by default
The PPv2 driver allocate lots of DMA buffers. These buffers are allocated
thanks to CMA if CONFIG_DMA_CMA is enabled, the size of the DMA pool is
configurable and too small in the default configuration.
This lead to the driver failing to probe because of this.
One solution is too disable CONFIG_DMA_CMA.
This patch might be reverted when backport patches for better solution.

Change-Id: I539b2f2347171c0217d52e52653715fe8dac4dff
Signed-off-by: Antoine Tenart <antoine.tenart@free-electrons.com>