arm64: dvfs: armada-37xx: fix cpu stall issue in switching from L2/L3 to L0 of 1200MHZ
Now for CPU L0 frequency @ 1.2GHz, the switching from L2/L3 to it
causes CPU stall. The reason is that VDD voltage has a big step in
switching from L2/L3(200/300MHz) to L0(1.2GHZ), , which requires
the time for stabling VDD voltage, but the VDD change latency
can not be covered by the relevant HW countdown register.

This patch provides a WA as below:
1. First switch from L2/L3(200/300MHz) to L1(600MHZ);
2. Sleep 20ms for stabling VDD voltage;
3. Then switch from L1(600MHZ) to L0(1200Mhz).

Change-Id: I1e782849ece278ea4d2ab86d95419d9bfc43b808
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/46516
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>