fix: net: mv88e6xxx: modified default DSA tag mode
DSA_TAG is 4 byte marvell tag where as EDSA_TAG is 8 byte with
4 bytes user programable ether type and 4 bytes DSA tag.
Updated default CPU tag mode to DSA_TAG instead of EDSA_TAG.
Because, this is the most commonly used or recommended mode.

Note that egress mode of the port depends on dsa tag mode.
It must be set to EGRESS_UNMODIFIED in DSA_TAG mode, while
any other configuration is allowed in EDSA_TAG mode.

Change-Id: Ide59187100376549628a49ef3b0e739e52adb8e3
Signed-off-by: Ravindra Reddy K <ravindra@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/46890
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>