net: pp2x: fix: allocate tx descriptors zeroed
Reserved/unused fields in tx-descriptors must be ZERO.
Driver doesn't clear them in run-time to improve performance.
The descriptors are not static-variable but run-time allocated,
so the allocator must Z-alloc procedure.

The problem could be easy emulated on original code
by setting the allocated area with 0xA5 -- driver doesn't work.

Change-Id: I1971cd75151f9931f022e7a57fd7a2a09c0b1dba
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/47064
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Yuval Caduri <cyuval@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>