fix: clk: mvebu: cp110: always enable gate MG clock 5
Gate clock 5 together with gate clock 6 feed XSMI and CM3 for
KR-AP in CP110. If ppv2 ports on CP0 connected to XSMI on CP1
and there are no enabled ppv2 ports on CP1. XSMI won't be enabled.
This will cause system stuck during access to XSMI address space.
Also the CM3 clock should be enabled, otherwise it will cause
system stuck when CM3 starts the execution.

This is a workaround, the complete solution should be nesting all
clock providers and consumers in the CP110 driver.

Change-Id: I032a8d18d705fa3e1402e330f4d11b90e56c2ae6
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/47119
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>