net: mvpp2: allow to assign a reserved memory region for DMA allocations
This patch adds the possibility to reserve a DMA region in the device
tree, and to assign this memory region to the PPv2 device for DMA
allocations. This helps to avoid cases where the default CMA pool is to
small in the kernel, by having a PPv2 specific pool we can control.

Change-Id: I504bec6907681db119132c7cd6fc20ff27bea306
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52757
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>