fix: dts: a8k: sdhci: add vqmmc regulator support
On A8K DB CP0, the vqmmc volatage is controlled by GPIO,
and fixed to be 3.3V.
According to SDIO spec, the vqmmc should be 1.8V to
support eMMC HS200 mode.
The patch added the vqmmc regulator which sets the default
voltage to 3.3V and changes to 1.8V in eMMC HS200 mode.

Change-Id: I300d11506961fcb2f56dd4104678a6c631c2364d
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/47633
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>