fix: arm64: clk: ap806: initialize the cluster clock by first valid CPU
During current AP806 cluster clock initialization, the cluster clock is
fixed to be initialized by the first CPU in despite of whether it is
valid or not, which makes the cluster clock uninitialized in case the
first CPU within the cluster is disabled, while the other CPUs within
this cluster are still valid, finally these CPUs will fail to get the
cluter clock and lead to issue.
For example, when disable the third CPU in u-boot environment, the
second cluster clock will not be initialized, which leads to the
failure in DFS probe routine when the fourth CPU get the clock,
finnally DFS fails to register "cpufreq-dt" type of platform device.
In order to fix this issue, need to initialize the cluster clock
by the first valid CPU.

Change-Id: I2ff8904a919854236dc87bb91c0b4aa5081aec82
Signed-off-by: Victor Gu <xigu@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/47740
Reviewed-by: Omri Itach <omrii@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Tested-by: Hanna Hawa <hannah@marvell.com>