dts: a8k: mcbin: enable pcie serdes init in kernel
Since the A8K Macchiatobin board has the EP reset
connected to GPIO, which is needed by the PCIe serdes
power up/down sequence from HW guidelines, so it is
necessary to support PCIe serdes initialization in kernel.

To initialize serdes in kernel, it need to power off
serdes first to remove dependency on bootloader, and
then implement the PCIe serdes power up/down sequence
with COMPHY lanes that is binded in DTS.

There is a property of skip_pcie_power_off to skip PCIe
serdes power off in kernel and keep the configuration from
bootloader, so for Macchiatobin board the property should
be removed. For other boards, such as A7/8K-DB, the
property is still retained and depend on bootloader
to initialize PCIe serdes.

The patch remvoed the property of skip_pcie_power_off
for A8K Macchiatobin.

Change-Id: I9614682802c64b3488e582164f12210f09c87b93
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/48119
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>