arm64: ahci: mvebu: Turn off LED when SATA link is offline
MVEBU AHCI has supported software activity led blinking, but if
the led blink policy is configured as BLINK_OFF(the LED is on when
idle, and blinks off every 10ms when activity is detected) and the
sata disk is plugged out, the led will be still in on state.

This patch overwrites the ata port postreset function which is called
after SATA link up/down events to support SATA LED present function
with ahci_mvebu_postreset(); besides the generic postreset operation,
ahci_mvebu_postreset() turns off LED when SATA link is offline.

Change-Id: If3b7948dece8a6b256f7d10bc1707106e182843b
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/48622
Reviewed-by: Ofer Heifetz <oferh@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>