arm64: dts: marvell: add vqmmc regulator on a7k DB for UHS
On A7K DB CP0, for some SD cards they can not be detected
or detected information wrong. The reason is that for UHS
card vqmmc should be 1.8V and without vqmmc regulator the
voltage is fixed to 3.3v and will fail to swicth to UHS
mode.

The patch added vqmmc regulator for A7K boards and updated
the 'armada7040-db-setup.txt' with the GPIO configuration.

Change-Id: I15fb14dde394daa078ee0ba30bed7640d0b627b7
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/48843
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>