net: mvpp2: align aggregated queue struct to cache
This patch is FIX and Optimization.

The Aggregated TX queue descriptors are allocated and
referenced as array of descriptors/structures "per CPU".
To avoid "cache line interference " on array's entry
the descriptor address and size should be be aligned to the
cache-line-size.
This alignment also improves the performance since 1 cache-line
used by CPU instead of 2 cache-lines.

Change-Id: I63aacb7f8a88bf26a3acc821b60beefc889d2286
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/49254
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52773
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>