net: mvpp2: Simplify MAC filtering function parameters
The mvpp2_prs_mac_da_accept function takes into parameter both the
struct representing the controller and the port id. This is meaningful
when we want to create TCAM entries for non-initialized ports, but in
this case we expect the port to be initialized before starting adding or
removing MAC addresses to the per-port filter.

This commit changes the function so that it takes struct mvpp2_port as
a parameter instead.

Change-Id: I014b0b66067504f004b897f9f88df28e0e749e80
Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52759
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>