crypto: inside-secure - improve the counter computation
A counter is given to the engine when finishing hash computation. It
currently uses the blocksize while it counts the number of 64 bytes
blocks given to the engine. This works well for all algorithms so far,
as SHA1, SHA224 and SHA256 all have a blocksize of 64 bytes, but others
algorithms such as SHA512 wouldn't work.

Change-Id: I2fd6a5a17a1934d45b5b9e57089e48bc4b93ee85
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/55725
Tested-by: iSoC Platform CI <ykjenk@marvell.com>