net: mvpp2: fix 10G port start stop dev sequence on phylink
PROBLEM: port 10G doesn't work if start_dev or/and stop_dev
is applied under ingerss traffic (which is in a life network).
The same problem is also on bootup sice it includes the start_dev.

SOLUTION:
There are 3 HW requirements on 10G ports XLG registers:
1). start_dev: The XLG configuration should be applied under reset.
2). start_dev: The configuration should enable XLG Control4 bit for
    hw-checking of 3 consecutive IDLEs before UP indication.
3). stop_dev:  The port-disable and port-reset should be done in
  2 separated steps with small delay (regiter-read is enought).

NOTE: This is re-work after PHYLINK support introducing.

Change-Id: Ie3ceb00f2050939c82282cdf97c7d877522774a0
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/54677
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>