crypto: inside-secure: move the digest to the request context
This patch moves the digest information from the transformation context
to the request context. This fixes cases where HMAC init functions were
called and override the digest value for a short period of time, as the
HMAC init functions call the SHA init one which reset the value. This
lead to a small percentage of HMAC being incorrectly computed under
heavy load.

Change-Id: Ie8ec54808ca80edabc02812a2bdc94d12bf5200c
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-by: Antoine Tenart <antoine.tenart@bootlin.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52003
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Nadav Haklai <nadavh@marvell.com>