cp110: pcie: set preset request vector to 6 presets (3-8)
Presets are used during the initial part of the equalization master phase.
Currently we use only preset4 in Linux and 6 presets (3-8) in u-boot.

The benefit of using more presets is that with different presets the PHY
can handle different type of interconnect, improving our support for
different types of hosts.

This patch aligns Linux presets configuration to u-boots and sets
6 presets (preset3-8).

Change-Id: I37e1f9c4869d769a41e0853117dcbeb6fa4eb68b
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52657
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>