fix: a7k: adjust AP806 SDIO regulator per board
Currently the SDIO regulator of AP806 is put in dtsi file
of "armada-ap806.dtsi" with part properties and its
status is okay by default, so for any board whose dts
file include "armada-ap806.dtsi", the SDIO regulator
is always valid, which cause the warning issue on the
board without SDIO enabled, such as A3900 board.

From point view of dts orginization, "armada-ap806.dtsi"
is in SoC level, and SDIO regulator should not be put
in the file of this kind, which should be put in board
level dts file, such as "xxx-db.dtsi" or "xxx.dts".

So the patch put AP806 SDIO regulator of "sd_mmc_reg0"
with full properties in dts file of board level and
fixed the warning issue on board without SDIO enabled.

Change-Id: I0afc2e2d3cfbf275bae458c5e9694d7e7504b1d3
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/52705
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>