fix: mvebu: icu: use syscore_ops instead of cpu notifier for PM
Currently we use CPU notifier to prepare ICU driver for
system power down. This a bug when CPUIDLE is enabled,
because CPU notifier is called not only for system power down
but for cpu power down too, causing endless store/restore of
the ICU state which might end up with ICU running with
incorrect state (if one of the CPU updates it during power-down,
then, ICU state is changed and the incorrect state is restored
during power-up.

This patch changes ICU driver to use syscore_ops instead of
the CPU notifier, achieving two benefits:
- Fixing the above issue, when CPU idle is enabled.
- There's no justification to use CPU notifies in the ICU driver
  because the ICU state is representing system view of interrupts
  assignments from CPs, and is irrelevant to CPU state, thus system
  notified syscore_ops is exactly what is required.

Change-Id: I9986dd2b8b9310910a0a7429342bda60f0a0b445
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53338
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>