net: mvpp2: Fix TCAM filter reserved range
This commit fixes the MVPP2_PE_LAST_FREE_TID define
that sets the end of the "free range", which included the MAC range.
This could therefore allow some other features to use entries dedicated
to MAC filtering, lowering the number of unicast/multicast addresses
that could be allowed before switching to promiscuous mode.

Change-Id: I57dafa059faf30f8b574161659ba73a21aca0d76
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53393
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>