pcie: armada8k: add support for comphy initialization
On armada8k platforms the comphy (common phy) must be configured to work
in appropriate mode. Hitherto comphy configuration for pcie was done by
bootloader. As a comphy driver was extended to support pcie comphy
configuration, this patch uses it in pcie driver. Thanks to that there
is no need to relay on the bootloader configuration which could
configured the comphy to different mode.

Change-Id: Icadd2739f048b454ce113b4e1921b804cb5376c2
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53856
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>