pinctrl: mvebu: ap806/cp110: remove hard coded gpio ranges
Since gpio ranges have been set in fdt gpio bank nodes for armada 7k
and 8k, so this patch removes hard coded gpio ranges for ap806 and
cp110.

And the hard coded gpio ranges are wrong for cp110-0 of armada 80x0,
only gpio bank 1 is used and the gpio range is from 52 to 82, this
patch also fixes the bug.

Change-Id: I0d94f35d5cb61a7df07bd0d7c82a8799210c5154
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53472
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: Omri Itach <omrii@marvell.com>