ata: libahci: set phy mode to SATA before powering it on
Some phy drivers like "marvell,comphy-cp110" are able to configure phy to
various mode, not only particular one. In such case it is required to set
appropriate mode before powering the phy.

For other phys which are only dedicated for SATA the set_mode handler is
not used therefore this patch shouldn't change anything.

Change-Id: Ie047fb0dc0f1d8161eede09c92dd8a14179688ae
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53861
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>