usb: hcd: set phy mode to usb host before powering it on
Some phy drivers like "marvell,comphy-cp110" are able to configure phy to
various mode, not only particular one. In such case it is required to set
appropriate mode before powering the phy.

For other phys which are only dedicated for USB host the set_mode handler
is not used therefore this patch shouldn't change anything.

Change-Id: Idd12e105e82f8250fcad08b5502ad724f16d61b3
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53864
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>