usb: hcd: set phy mode to usb host before powering it on
Some phy drivers like "marvell,comphy-cp110" are able to configure phy to
various mode, not only particular one. In such case it is required to set
appropriate mode before powering the phy.

For other phys which are only dedicated for USB host the set_mode handler
is not used therefore this patch shouldn't change anything.

Change-Id: I04c034ec736ac0b5149c6f3103cff95773493732
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>