fix: pcie: aardvark: correct way of PIO status checking
There is an issue that when PCIe switch is connected
to Armada 3700 board, there will be lots of warning PIO
errors when reading config space.
According to Aardvark PIO read and write sequence in
HW specification, the current way to check PIO status
has following issues:
1) For PIO read operation, it reports the error message,
   which should be avoided according to HW specification.
2) For PIO read and write operations, it only check PIO
   operation complete status, which is not enough and
   error status should be checked.

The patch aligned with Aardvark PIO read and write sequence
in HW specification on PIO status check and the warnning
errors when reading config space.

Change-Id: I83a755e95a518164978361ea703409accfafcc09
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/53739
Reviewed-by: Victor Gu <xigu@marvell.com>
Tested-by: Victor Gu <xigu@marvell.com>