net: phy: sfp: warn the user when no tx_disable pin is available
In case no Tx disable pin is available the SFP modules will always be
emitting. This could be an issue when using modules using laser as their
light source as we would have no way to disable it when the fiber is
removed. This patch adds a warning when registering an SFP cage which do
not have its tx_disable pin wired or available.

Change-Id: Ia8407dab981f7797956601480887086bef0a6640
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/54578
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>