fix: thermal: armada: register DT thermal sensor using device data
To reduce compatible comparison for each sensor that supports
DT thermal zone, added a new boolean flag to device data to indicate
support for DT sensor.

After this change the registeration is done according to this
flag.

This is a preparation patch for new SoCs to be added to this driver.

Change-Id: Ifd09c2be7649622f3883ad9fc3b2d89fc568d057
Signed-off-by: Christine Gharzuzi <chrisg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/54701
Reviewed-by: Omri Itach <omrii@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>