dts: armada-37xx: reserve 16M memory for tee
These memory are used for TEE code, data, communication channel
and shared memory with Non-Trusted world.

For the TEE code and data section, it will be configured
unacceable to Non-Trusted World by memory controller.

The rest memories are either used for communication or shared
memory between Trusted-World and Non-trusted World, which is
TEE implementation defined and may be managed by the TEE driver.

Change-Id: Ie60b7a65d3a5c845cf51f68797130051ff39f659
Signed-off-by: wangwen <wangwen@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/55267
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Kostya Porotchkin <kostap@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>