crypto: inside-secure - split the algorithms list by engine version
This patch splits the common algorithms list by engine version, and puts
it in the device data. Minor reworks are also done as the device data
was already used to store information about the version of the engine
used. This is needed for sha384/512 support as those algorithms will be
supported by the eip197, but not the eip97.

Change-Id: Ibf7c5fb2d2e3c017c87796b859b179208155e5a3
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>
Signed-off-by: Ofer Heifetz <oferh@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/55488
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>