serial: 8250: set RCVR trigger to '1 character in the FIFO'
Currently, there's an overrun issue with serial interface on
Armada8K devices (mainly with MCBin).
When pasting long lines to Linux promt, overrun might occur.
This happens because characters arriving to the FIFO faster then
they're handled by the CPU.

This patch changes "RCVR Trigger" settings, from "FIFO ½ full" to
"1 character in the FIFO". This modification triggers the RX
interrupt faster, which clears the FIFO faster,
avoiding the overruns described above.

Note, this is a temporary fix, to avoid the overrun issue,
more generic fix is required for mainline.

Change-Id: I90abb7f013ce138da04b802461056b32cf2a7b0f
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/55907
Reviewed-by: Yoav Gvili <ygvili@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>