dts: a8kp: remove pcie comphy handlers and set skip_pcie_power_off
Be default, in a8kp development board, the PCIe reset is connected
to the system reset so PCIe reset happens once, when the SoC boots.
We can't reset PCIe per request - this implies that we cannot run the
complete PCIe power-up sequence, so Linux must rely on bootloader
PCIe comphy configuration.

This patch does the following:
- Add skip_pcie_power_off property to comphy node:
  This allows us to slip PCIe power-off and keep the configuration done
  by the bootloader.
- Remove PCIe comphy handlers:
  This allows to avoid execution of power-on sequence during Linux boot.

Change-Id: I859e3503c87637d7939cc1696dc789a5a0755243
Signed-off-by: Igal Liberman <igall@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/55929
Reviewed-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>