arm64: dts: marvell: armada-cp110: change the PPv2 IRQ names
This patch changes the PPv2 IRQ names in the CP110 device tree to match
a corresponding change in the Marvell PPv2 driver. The reason this was
updated is the IRQ where names after Tx/Rx interrupts, but this can be
configured and ithose interrupts are not mapped to a given queue. A
following patch will add more of them and the names wouldn't make sense.

Change-Id: I5a9c99981ff9d66748457423d1bffa0e4be379e8
Signed-off-by: Antoine Tenart <antoine.tenart@bootlin.com>