fix: net: pp2x: remove unused buff-header handler
On link down under active traffic the RX-status word in the
RX-descriptor may have wrong <buffer header bit>=set.
This leads to SKB=NULL fetching from descriptor and PANIC.

FIX:
The buff-header (Scatter/gather) is not really supported is removed.
So even if the bit <buffer header bit> is set the bad-packet would
be detected and passed correctly over err-drop.

Change-Id: I144164e35a99dfc6e94f20c7fe1477dbb12ad344
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/56216
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>