coresight: axim: Added driver for the AXI monitor
The AXI monitors is a bus monitor that exists in both AP-806 and
CP-110. it is capable of detecting specific bus transactions and
generate trace events or increment counters

This driver adds the basic support for the AXI monitor. it enables
using the AXI monitor for debug purposes. future patches will add
support for tracing using the axi monitor

Documentation/trace/axi-monitor.txt provides a basic explanation on
the monitor on how to use this driver

Documentation/ABI/testing/sysfs-bus-coresight-devices-axim gives
detailed descirption of all the sysfs file needed to operate the AXI
monitor

Change-Id: Iea4d71ab8fc2c9db5ad25bd5763ba43359dddd94
Signed-off-by: Omri Itach <omrii@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/56909
Reviewed-by: Igal Liberman <igall@marvell.com>
Tested-by: Igal Liberman <igall@marvell.com>