vfio: xhci: add support for XHCI reset hook
The VFIO reset hook is called every time a platform device is
passed to a guest or removed from a guest.
When the XHCI device is unbound from the host, the host driver
disables the XHCI clocks/phys/regulators so when the device is
passed to the guest it becomes dis-functional.
This initial implementation uses the VFIO reset hook to enable
the XHCI clocks/phys/regulators on behalf of the guest.

Change-Id: I4b88e4b79889fc6067fb3f33753b5b4a61b5f157
Signed-off-by: Yehuda Yitschak <yehuday@marvell.com>
Signed-off-by: David Sniatkiwicz <davidsn@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/58162
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hanna Hawa <hannah@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>