arm64: dts: marvell: add current-limiter and vbus regulator to USB3 in a8kp DB
USB vbus can be configured through the io-expanders found on the board. The
expander0 is responsible for configuring USBs from CP0.

Before this patch, the current-limiter regulator relied on U-Boot proper
setting for 900mA. this patch removed this dependency by fixing the
regulator settings to properly use 900mA when needed.

- The USB Host should be linked to current-limiter through
  "current-limiter-supply" DTS property.
- This patch adds the "regulator-fixed" node for the USB current limiter.
- This patch adds handling for "current-limiter-supply" property inside
  the USB3 node in armada-8082 device trees for turning on
  the current limiter

Change-Id: Icbd6d5732a90ebdcfdbafeabd9c9bd2b719b3e84
Signed-off-by: Christine Gharzuzi <chrisg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57577
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>