fix: dma: mv_xor_v2: configure timer threshold after getting interrupt
Issue found while working with CPU@100Mhz, the XOR engine sends lot of
interrupts.
The XOR engine threshold feature sends interrupts after X time
passed if no one handle the ready describtors, and this creates lot
of interrupts and tasklets in the queue.

This patch increase the timer threshold to be maximum when handle the
interrupt and configure it back to default when handle all the ready
describtors.

Change-Id: I7eb6b801bd1f30715da1692a4f525ee12b90e95e
Signed-off-by: Hanna Hawa <hannah@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57698
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Ofer Heifetz <oferh@marvell.com>