net: mvpp2: add tx bulk timer needed for performance optim
The traffic performance could be improved (up to 22%) by reducing
an access to Aggregated TX-queue counter register
(write a bulk of N instead of one-by-one).
This requires defer in xmit-procedure and bulk-timer.

This patch introduces infrastructure for the feature
(structures, timer, tasklet, initialization), but does not modify
the xmit. Since the timer-start is not called by xmit the feature
is still disabled.

Change-Id: I4aafa3db3813ab843b61a992880b05a92261ddcc
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/57746
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Stefan Chulski <stefanc@marvell.com>
Reviewed-by: Igal Liberman <igall@marvell.com>