net: pp2x: check skb-data-buffer is not replaced by NET-stack
When a RXed skb with its dtata-buffer passes over NET-stack or
NET-application the same skb may be reused but with ANOTHER data-buffer.
Since the size of this data-buffer is not guaranteed it should Not
be decided as recyclable.

We have a MAGIC generated from skb-address and kept in the same skb
in skb->cb[].
Let's use not the skb address but data-buffer address skb->head
for MAGIC-generation and checking. So if the data-buffer has been
replaced by NET, the MAGIC saved on RX and MAGIC calculated in TX
would be different and the buffer would be decided non-recyclable.

NOTE: the skb->data pointer may be moved up/down inside of the
data buffer but this does not makes the buffer non-recyclable.

Change-Id: Id3d43565cba72126b0e75f4151edfc51b631dcb7
Signed-off-by: Yan Markman <ymarkman@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/58621
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>