drivers: pci: controller: Add Octeon PEM driver
This driver adds support for link-down event in PEM
controller to support hot-plug removal of endpoint devices.
An interrupt handler is registered for RST_INT msix vector
which is triggered with link going down. This handler
performs cleanup of root bridge and its children and
re-initializes root bridge to kernel for next link-up event.

Change-Id: Ib9df8bf32c29badbf1762a16ce29c58db28753b4
Signed-off-by: Suneel Garapati <sgarapati@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/54416
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/96250
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>