net: thunderx: Fix transmit channel config for secondary
Due to the changes introduced for loopback (LBK) interface
support, VNIC's secondary qsets transmit path is broken. This
is due to the incorrect way of checking LBK mapped VF which
resulted in configuring transmit channel as PKO instead of
CGX for secondary Qset SQs.

This patch fixes the issue, also cleaned up the LBK VF checks
throughout the PF driver.

Change-Id: Ic44d81ce4a29dbd6760ce8735a9be9b6de7cb49a
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Signed-off-by: Harman Kalra <hkalra@marvell.com>