octeontx2-af: suppress kpu profile loading warning
Currently there are three way in which a KPU profile can be
loaded (in high to low priority order):
1. profile image integrated in kernel image
2. firmware database method
3. default profile

In most of the cases profile is loaded using 2nd method, resulting
in a warning from linux firmware subsystem (method 1) due to absence
of firmware in kernel image.
Using an appropriate firmware_request API which can suppress such
warnings if no image is integrated to kernel image.

Signed-off-by: Harman Kalra <hkalra@marvell.com>
Change-Id: Ie9299ff70be1a24fa7d3a18249ab793e33b685e9