octeontx2-af: timeout while draining SSO queues
Draining SSO queues won't finish on CN10k hardware if XAQ gets
corrupted. So a 1 minute timeout is added for the drain process.

The XAQ corruption could be due to a double bit error and chance of it
happening are very rare. So this issue won't be fixed in silicon.

The timeout is added for all silicon as FLR times out in 64k ms.

Drain retry count is also reduced to move between queues quickly and
finish the drain faster.

Signed-off-by: Shijith Thotton <sthotton@marvell.com>
Change-Id: I678fbb1749188005abcc204fd837ce20c0c769a9
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/79613
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>