firmware: smccc: Add support for erratum discovery API
It is not always possible for the OS to determine if a CPU is affected by
a particular erratum. For example, it may depend on an integration choice
the chip designer made, or whether firmware has enabled some particular
feature.

In particular the 'TC' erratum can be worked around by setting an EL3
chicken bit. Knowing whether firmware has support for this erratum
workaround will all the kernel to configure it, and report the workaround
state to user-space.

Add support for the SMCCC 'Errata Management Firmware Interface' that lets
the OS query firmware for this information. This will allow KVM to
report the issue as mitigated for nested guests that run on these platforms.

Link: https://developer.arm.com/documentation/den0100/1-0/?lang=en
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Tanmay Jagdale <tanmay@marvell.com>