drivers: pci-octeon-pem: Rectify RC link recover work
In worker thread for RST_LINKDOWN,
first check if hot-plug interrupt thread is in progress
due to change in slot status bits.
Wait for it to complete and then disable Hot-plug interrupt.

This helps avoid any link events that trigger HP interrupt
thread in parallel leading to a deadlock.

Then proceed with clean up of Root bridge and children,
next is to check link status and if required toggle PEMON
to set hardware to correct state and then re-scan to setup
root bridge for next link-up event.

Finally, add dependency on HOTPLUG_PCI_PCIE config option
based on above changes.

Change-Id: I9af6c4f1381dc01dedb50c9a2a0a68eee68acc8d
Signed-off-by: Suneel Garapati <sgarapati@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/87412
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/96253