coresight: tmc: Add support for register metadata
Parse the previous boot register metadata details from
the device node. This register metadata is used to read
previous boot preserved trace.

On platforms that has provision for firmware intervention
in the watchdog reset path, firmwware is expected to sync
the metadata region with the coresight register snapshot
in the specified format.

Change-Id: I8fe5d66b6c61dfab11794d13feaf5ad7a27ea73d
Signed-off-by: Anil Kumar Reddy <areddy3@marvell.com>
Signed-off-by: Tanmay Jagdale <tanmay@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/102652
Tested-by: Linu Cherian <lcherian@marvell.com>
Reviewed-by: Linu Cherian <lcherian@marvell.com>