octeontx2-pf: Check for DMAC extraction support before setting DMAC based hardware filter for a VF
As of today, while setting a MAC address of a PF's VF (e.g. ip link set
<pf-netdev> vf 0 mac <mac-address>), it blindly tries to install a DMAC
based hardware filter. But it is possible that the hardware parser profile
does not support DMAC extraction. This patch fixes the same and tries
to add the DMAC filter only if hardware parsing profile supports DMAC
extraction.

Fixes: f0c2982aaf98 ("octeontx2-pf: Add support for SR-IOV management functions")
Signed-off-by: Suman Ghosh <sumang@marvell.com>