x86/resctrl: Add stub to allow other architecture to disable monitor overflow
Resctrl has an overflow handler that runs on each domain every second
to ensure that any overflow of the hardware counter is accounted for.

MPAM can have counters as large as 63 bits, in which case there is no
need to check for overflow.

To allow other architectures to disable this, add a helper that reports
whether counters can overflow.

Signed-off-by: James Morse <james.morse@arm.com>
Change-Id: I4963832326363aa1512c70545fb1100a1356f2b3
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/144050
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>