octeontx2-pf: Add AF_XDP zero copy support for rx side
This patch adds support to AF_XDP zero copy support for CN10K.
This patch specifically adds receive side support. In this approach once
a xdp program with zero copy support on a specific rx queue is enabled,
then that receive quse is disabled/detached from the existing kernel
queue and re-assigned to the umem memory.

Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I81845e3b4132eb23f43b2660d23e5997acc09a0b
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/130867
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/154836