fs/resctrl: Make rdtgroup_destroy_root() safe to call when not mounted
When resctrl_exit() is called, rdtgroup_destroy_root() is used to do
the opposite of rdtgroup_setup_root(), but this is only called when the
filesystem is mounted. If rdtgroup_destroy_root() is called when the
file-system is not mounted, kernfs_destroy_root() dereferences
a NULL pointer.

resctrl_exit() should take the rdtgroup_mutex to avoid raceing with
rdtgroup_setup_root(), and rdtgroup_destroy_root() should check whether
rdt_root is NULL.

Signed-off-by: James Morse <james.morse@arm.com>