net: mvpp2: Fix tag mode for DSA switch connection
The DSA tag mode is disabled by default. When a port is connected to the
DSA switch, the Parser rules for matching non-extended DSA tag should be
enabled so that the L3 & L4 info could be updated properly in RI and be
conveyed to Classifier for further processing.

All the mvpp2 ports are registered to the netdev notifier. When a DSA
port is linking to the mvpp2 port, switch the tag mode to non-extended
DSA tag as well as enable DSA pre-classifier.

Signed-off-by: Wilson Ding <dingwei@marvell.com>
Change-Id: I7a733088b328e56085033e32f2a22cbadf517ab6
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/149605
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sanghoon Lee <salee@marvell.com>