arm64: dts: marvell: Clean redundant property for cn9130-crb-r1p3-*.dts
The MPP configuration of CP0 SPI1 is aligned with CRB-CN9130 boards of
R1P3 revision when CRB board was introduced to Linux. Unlike SDK10 or
SDK11, we are not going to support the boards which the revisions are
earlier than R1P3. Hence, remove the pinctrl property of cp0_spi1 in all
cn9130-crb-r1p3-*.dts.

Signed-off-by: Wilson Ding <dingwei@marvell.com>
Change-Id: Ibf12b13584702e0db14a9a68aad635e8d6e5bb06
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/149633
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sanghoon Lee <salee@marvell.com>