octeontx-bphy-netdev: Fix unexpected packet in netdev by ensuring proper PSW hardware write handling
Checks if the PSW length is zero in cnf10k_rfoe_process_rx_pkt().
If the PSW length is zero, defers processing of the mbt_index
to the next cycle.
Additionally, memset the PSW address in every cycle
to ensure proper handling.
If the PSW length is zero, poll napi_poll again by returning
the budget in this condition.
This allows the current mbt_index to be processed in the next cycle.

Signed-off-by: Viswajith Murali <viswajithm@marvell.com>
Change-Id: I4bbde4481b1b73f18abd2a7804429a0f9d95eb8e
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/152201
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>