octeontx2-af: Fix transmit timeout errors
As part of the link mode change process, the TL1 SW_XOFF  is
enabled before sending the command to the firmware. This needs
to be cleared  when a LINK event is received.

On CN20K, the TL1 SW_XOFF remains set even after a LINK_UP event,
which leads to transmit timeout issues.

This patch fixes the same.
Fixes: 55e39e33ccd8 ("octeontx2-af: poll for tx link credits before link mode change")
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Change-Id: Id537b2873a96fcec60e5d90f4bf230fc671dd4a6
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/152959
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>