net: mvpp2: Add support for eBridge DSA tag
eDSA can be used shorten for both eBridge DSA and extended DSA. The only
difference is that eBridge DSA is 16B while extended DSA is 8B.

Since both DSA tag are not supposed to be enabled simutaneously. The
eBridge DSA support reuse the Parser entries pre-allocated for extended
DSA.

To simplify the Parser lookup for eBridge DSA, it bypasses the VLAN
filtering which is supposed to be guaranteed by the DSA switch.

Signed-off-by: Wilson Ding <dingwei@marvell.com>
Change-Id: I9f0106f1cd7455b3f3885792d2b94f396f848d3b
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/153484
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sanghoon Lee <salee@marvell.com>