octeontx2-af: Fix PF driver crash with kexec kernel booting
When both AF and PF drivers are built as modules, the PF driver in the
kexec kernel may probe before the AF driver is ready. This leads to
a crash due to uninitialized hardware state.

This patch ensures the PF driver properly detects and waits for AF
driver readiness before proceeding with initialization.

Fixes: 54494aa5d1e6 ("octeontx2-af: Add Marvell OcteonTX2 RVU AF driver")
Signed-off-by: Anshumali Gaur <agaur@marvell.com>
Change-Id: I7fef7198af3631aef8f4478ad4d7ef83b0463f81
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/157334
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
Reviewed-by: Geethasowjanya Akula <gakula@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Sai Krishna <saikrishnag@marvell.com>