pci: reduce baseband PHY device BAR0 size
The baseband PHY device on Marvell's CN9K and CN10K
silicon features a BAR0 of 512GB, which consumes a
significant amount of memory when mapped with vfio-pci
using a 4K page size.

This patch reduces the BAR0 size to 129GB, since no
hardware resources are mapped beyond this limit.

Signed-off-by: Vamsi Attunuru <vattunuru@marvell.com>
Change-Id: I09c8216c491a47dae093174f2930ea9f09fbc687
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/157646
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>