octeontx: pan: Add support for KTLS Tx offload
This patch provides supports for KTLS Tx offload. But there are few
limitation to it and will be supported by incremental patches. Below are
the limitations:
(1) TLS Record over the MTU size where TLS reassembly or CPT offload for
 individual TCP segment is required is not supported.
(2) TCP out-of-order scenario is not supported.
(3) Currently there is a hard coding of channel number for eth0 and
sdp-1, which needs to be removed after finalizing the API between
pan_switch and pan_ktls

This patch is working for TLS records which is <= (MTU size - ETH + IP +
TCP header). Also, this patch is having support for mailbox
communication between oceton_ep host driver and pan.
The intention of this patch is to add basic infra required for KTLS
offload. In short below are the things that the patch support:
(1) Handle add/del of ktls context from octeon_ep host to pan
(2) Add NPC rule along with the pan context
(3) Does CPT offload of the KTLS unencrypted packet in the datapath

Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I200d989e2802c735c305064f31483b34442323a2
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/158468
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>