octeontx2-vf: Set tc filter flag on vf netdev
Once a vf device is created, by default no mcam entries are allocated fo
that netdev. It can be done by calling devlink command "mcam_count". But
that call was only setting the ntuple flag on the netdev. Because of
that only ethtool ntuple filter was working not the tc filters.
This patch fixes the issue by setting the tc filter flag during
the mcam entry allocation.

Fixes: 2da489432747 ("octeontx2-pf: devlink params support to set mcam entry count")
Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I3174731bce842d45e55588d765bdbf98fdf74a68
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/159873
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>