octeontx: pan: Fix SNAT with routing
SNAT won't be supported in "routing table based L3"
in PAN. we need to get into flows to understand the
packet to masquerade properly. Linux nftables has HW
offloading support (only for L3). we leveraged the frame
work to push L3 flows. we can extend to enable various L3
forwarding and magling. For now, we have only L3 forward with
 SNAT and DNAT in PAN.

Signed-off-by: Ratheesh Kannoth <rkannoth@marvell.com>
Change-Id: I53fb72e762257a379bcbb5a734938154118acc2b
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/160656
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>