SDP: Broadcast XON on SDP channels
On the CN106 silicon, the SDP_LINK_CFG is configured by the ATF,
but on all other silicons, it's handled by the kernel's SDP driver.

The previous patch fixed the "SDP channel XOFF" issue specifically
for CN106 silicon, but the problem persists on other silicons.
The workaround only works if the broadcast message is sent after
the SDP_LINK_CFG  register is configured.

This patch resolves the issue by ensuring that the workaround is
applied only after the SDP_LINK_CFG register has been properly configured.

Fixes: 98a1cc39b439 ("Octeontx2-af: SDP: Broadcast XON on SDP channels")
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Change-Id: I68e96f289b937c8695eb32c5c17d80aa765e04d0
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/162253
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>