of: Add device only if associated CPU is online
When we boot the SoC with fewer cores than physically available
via the "maxcpus" kernel command line argument, the per core
AMBA devices such as ETF, PMPCSR, CTI etc fail to probe when
their associated CPU is offline and results in RAS errors.

Hence, to avoid this error, add a check in of_amba_device_create()
routine to check if the device's associated CPU is online and
skip it if offline.

Change-Id: I0192ead6629fcd2e7585f10cb6216a8cd9240e08
Signed-off-by: Tanmay Jagdale <tanmay@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/164913
Reviewed-by: Devapraba Muthumani <dmuthumani@marvell.com>
Tested-by: Devapraba Muthumani <dmuthumani@marvell.com>