octeontx2-sdp: Add FLR interrupts for Host PFs
When a host PF FLR is initiated then all the rings alloted
to that host goes to FLR and interrupts are raised to SDP.
There are 8 interrupt vectors and registers indicating all
512 rings of SDP. This patch adds FLR handling support by registering
interrupts for 8 vectors. To simplify FLR handling, if an interrupt
out of 8 is raised then all the FLR-ed rings of all PF/VFs are
checked and cleaned. Also futher interrupts are disabled during tear down
process to avoid race conditions.

Change-Id: I73d6a64c11af0fb24238f4dbdc1292e27a683347
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/165784
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>