octeontx2-af: Wrap CN20K debugfs code with CONFIG_DEBUG_FS
The CN20K AF debugfs support was always compiled, even when
CONFIG_DEBUG_FS was disabled. This resulted in compilation
errors as it refers to the structure rvu_dbg which is define
under CONFIG_DEBUG_FS.

Wrap the entire debugfs implementation in debugfs.c with
CONFIG_DEBUG_FS to ensure the file is only compiled when
debugfs support is available.

Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Change-Id: Iad129c9e89f978b4f8d271caa0b59912b1f35e2d
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/167230
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>