ras:cn10k:BCP issue fix where CE errors are read as UE in kernel
octeontx_ghes_record is a structure (which has some structures
and few variables as members).The member structure is modified
in new kernel(6.1) due to RAS changes(new firmware)in ATF to
report arm context info for external abort and core errors.
Hence the offset of variable "error_severity" got changed.

When old firmware is tested with new kernel,kernel is looking at
the error_severity at a different offset.

To fix the BCP issue, created a new structure below the variables.
This makes sure the existing structure and variables stay at the
same offset as old firmware.

Signed-off-by: Ashwin Kumar Sundararamakrishnan <asundararama@marvell.com>
Change-Id: I1eb96f39dd54a82dcda759cd017eb79f7e82ccda
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/167635
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>