CNF20K: CPRI: Add support for CPRI GPINT handling
This driver adds support for handling GPINT interrupts and provides a
callback mechanism for upper layers to receive GPINT notifications.

Unlike the Loki and Thor platforms, GPINT resources are virtualized
through the BGEN block, so a fixed GPINT resource cannot be used.
Each chiplet exposes a BGEN LF that is associated with the CPRI PF
device, and this mapping is conveyed to the kernel CPRI PF driver
via an ioctl interface.

The CPRI PF driver then registers a platform device that describes the
CPRI resources, including the CPRI base address and the GPINT vector
assigned to that chiplet.

Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Change-Id: Id417dcd9ac6a1917626c21653cb67a62bd1134b1
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/167780
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>